+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
 Since version 1.0.0, we try to follow the [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard.
 
 ## [Unreleased]
+
+## [1.4.0] - 2021-08-16
 ### Added
 
 - The plugin will now log an error if the output format of Zigbee2MQTT (`experimental.output`) appears to have been configured incorrectly.
@@ -229,7 +231,8 @@ For `cover` devices the following changes/fixes are in this release:
 - Improve state determination for WindowCovering.
 
 
-[Unreleased]: https://github.com/itavero/homebridge-z2m/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/itavero/homebridge-z2m/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/itavero/homebridge-z2m/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/itavero/homebridge-z2m/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/itavero/homebridge-z2m/compare/v1.1.3...v1.2.0
 [1.1.3]: https://github.com/itavero/homebridge-z2m/compare/v1.1.2...v1.1.3